Fix listSchemasLike when multiple Iglu repositories host the schemas
We had a problem with listSchemasLike when using the mustIncludeKey
option, which was introduced in #215

For a particular schema...

- Iglu Server 1 hosted versions `1-0-0` and `1-0-1`
- Iglu Server 2 hosted versions `1-0-0`, `1-0-1`, `1-0-2` and `1-0-3`

and `mustIncludeKey` was set to `1-0-3`. But `listSchemasLike` returned
only `1-0-0` and `1-0-1` because Iglu Server 1 had higher priority.

After this change, `listSchemasLike` will return the list from Iglu
Server 2 under these circumstances. Even though it has lower priority,
it is the server that contains the `mustIncludeKey`.