Add caching to lookupSchemasUntil function
Also, lookupSchemasUntilResult function is implemented as well. It is the variant of lookupSchemasUntil
that returns ResolverResult which contains information about cahe item.

Also, mimaPreviousVersions is cleared out because this commit breaks binary compatibility with previous versions
since it adds a new function to a trait. Therefore, we needed to clear-out the mimaPreviousVersions.